// src/graphql/graphql.ts
export type Maybe<T> = T | null
export type InputMaybe<T> = Maybe<T>
export type Exact<T extends { [key: string]: unknown }> = {
  [K in keyof T]: T[K]
}
// ...
export type Scalars = {
  ID: { input: string; output: string }
  String: { input: string; output: string }
  Boolean: { input: boolean; output: boolean }
  Int: { input: number; output: number }
  Float: { input: number; output: number }
  /** A date-time string at UTC, such as 2007-12-03T10:15:30Z, compliant with the `date-time` format outlined in section 5.6 of the RFC 3339 profile of the ISO 8601 standard for representation of dates and times using the Gregorian calendar. */
  DateTime: { input: any; output: any }
  /** A field whose value conforms with the standard mongodb object Id as described here: https://docs.mongodb.com/manual/reference/method/ObjectId/#ObjectId. Example: 5e5677d71bdc2ae76344968c */
  ObjectId: { input: any; output: any }
}
// ...
export type AudioUrls = {
  __typename?: 'AudioUrls'
  /** Female version of audio url of the post. */
  female?: Maybe<Scalars['String']['output']>
  /** Male version of audio url of the post. */
  male?: Maybe<Scalars['String']['output']>
}
// ...
export type Node = {
  /** The ID of the node. */
  id: Scalars['ID']['output']
}
// ...
export type Post = Node & {
  __typename?: 'Post'
  /** Returns male and female audio url of the post. Available in case the Audioblog is enabled. */
  audioUrls?: Maybe<AudioUrls>
  /** Returns the user details of the author of the post. */
  author: User
  /**
   * Flag to indicate if the post is bookmarked by the requesting user.
   *
   * Returns `false` if the user is not authenticated.
   */
  bookmarked: Scalars['Boolean']['output']
  /** Brief is a short description of the post extracted from the content of the post. It's 250 characters long sanitized string. */
  brief: Scalars['String']['output']
  /** Canonical URL set by author in case of republished posts. */
  canonicalUrl?: Maybe<Scalars['String']['output']>
  /**
   * Returns the user details of the co-authors of the post.
   * Hashnode users can add up to 4 co-authors as collaborators to their posts.
   * This functionality is limited to teams publication.
   */
  coAuthors?: Maybe<Array<User>>
  /** List of users who have commented on the post. */
  commenters: PostCommenterConnection
  /** A list of comments on the post. */
  comments: PostCommentConnection
  /** Content of the post. Contains HTML and Markdown version of the post content. */
  content: Content
  /**
   * A list of contributors of the post. Contributors are users who have commented or replied to the post.
   * @deprecated Will be removed on 10th Oct 2023. Use `commenters` instead.
   */
  contributors: Array<User>
  /** The cover image preference of the post. Contains cover image URL and other details. */
  coverImage?: Maybe<PostCoverImage>
  /** Unique ID to identify post, used internally by hashnode. */
  cuid?: Maybe<Scalars['String']['output']>
  /** Flag to indicate if the post is featured on Hashnode feed. */
  featured: Scalars['Boolean']['output']
  /** The date and time the post was featured. Used along with featured flag to determine if the post is featured. */
  featuredAt?: Maybe<Scalars['DateTime']['output']>
  /** Post feature-related fields. */
  features: PostFeatures
  /** A flag to indicate if the post contains LaTeX. Latex is used to write mathematical equations. */
  hasLatexInPost: Scalars['Boolean']['output']
  /** The ID of the post. Used to uniquely identify the post. */
  id: Scalars['ID']['output']
  /** Wether or not the post has automatically been published via RSS feed. */
  isAutoPublishedFromRSS: Scalars['Boolean']['output']
  /**
   * Wether or not the authenticated user is following this post.
   *
   * Returns `null` if the user is not authenticated.
   */
  isFollowed?: Maybe<Scalars['Boolean']['output']>
  /** A list of users who liked the post. */
  likedBy: PostLikerConnection
  /** OG meta-data of the post. Contains image url used in open graph meta tags. */
  ogMetaData?: Maybe<OpenGraphMetaData>
  /** Preference settings for the post. Contains information about if the post is pinned to blog, comments are disabled, etc. */
  preferences: PostPreferences
  /** The publication the post belongs to. */
  publication?: Maybe<Publication>
  /** The date and time the post was published. */
  publishedAt: Scalars['DateTime']['output']
  /** The number of hearts on the post. Shows how many users liked the post. */
  reactionCount: Scalars['Int']['output']
  /** The estimated time to read the post in minutes. */
  readTimeInMinutes: Scalars['Int']['output']
  /** The number of replies on the post. */
  replyCount: Scalars['Int']['output']
  /** The number of comments on the post. */
  responseCount: Scalars['Int']['output']
  /** SEO information of the post. Contains title and description used in meta tags. */
  seo?: Maybe<Seo>
  /** Information of the series the post belongs to. */
  series?: Maybe<Series>
  /** The slug of the post. Used as address of the post on blog. Example - https://johndoe.com/my-post-slug */
  slug: Scalars['String']['output']
  /** The subtitle of the post. Subtitle is a short description of the post which is also used in SEO if meta tags are not provided. */
  subtitle?: Maybe<Scalars['String']['output']>
  /** Returns list of tags added to the post. Contains tag id, name, slug, etc. */
  tags?: Maybe<Array<Tag>>
  /** The title of the post. */
  title: Scalars['String']['output']
  /** The date and time the post was last updated. */
  updatedAt?: Maybe<Scalars['DateTime']['output']>
  /** Complete URL of the post including the domain name. Example - https://johndoe.com/my-post-slug */
  url: Scalars['String']['output']
  /** The number of views on the post. Can be used to show the popularity of the post. */
  views: Scalars['Int']['output']
}
// ...
export type PostCoverImage = {
  __typename?: 'PostCoverImage'
  /** Provides attribution information for the cover image, if available. */
  attribution?: Maybe<Scalars['String']['output']>
  /** True if the image attribution should be hidden. */
  isAttributionHidden: Scalars['Boolean']['output']
  /** Indicates whether the cover image is in portrait orientation. */
  isPortrait: Scalars['Boolean']['output']
  /** The name of the photographer who captured the cover image. */
  photographer?: Maybe<Scalars['String']['output']>
  /** The URL of the cover image. */
  url: Scalars['String']['output']
}

/** Contains a post and a cursor for pagination. */
export type PostEdge = Edge & {
  __typename?: 'PostEdge'
  /** A cursor for use in pagination. */
  cursor: Scalars['String']['output']
  /** The node holding the Post information */
  node: Post
}
// ...
export type PublicationPostConnection = Connection & {
  __typename?: 'PublicationPostConnection'
  /** A list of edges containing Post information */
  edges: Array<PostEdge>
  /** Information for pagination in Post connection. */
  pageInfo: PageInfo
  /** The total number of documents in the connection. */
  totalDocuments: Scalars['Int']['output']
}
// ...
export type Query = {
  __typename?: 'Query'
  /**
   * Returns a draft by ID.
   * Draft is a post that is not published yet.
   */
  draft?: Maybe<Draft>
  /**
   * Returns a paginated list of posts based on the provided filter.
   * Used in Hashnode home feed.
   */
  feed: FeedPostConnection
  /** Returns the current authenticated user. Only available to the authenticated user. */
  me: MyUser
  /**
   * Returns the publication with the given ID or host.
   * User can pass anyone of them.
   */
  publication?: Maybe<Publication>
  /** Get a scheduled post by ID. */
  scheduledPost?: Maybe<ScheduledPost>
  /** Returns a paginated list of posts based on search query for a particular publication id. */
  searchPostsOfPublication: SearchPostConnection
  /** Returns tag details by its slug. */
  tag?: Maybe<Tag>
  /** Returns users who have most actively participated in discussions by commenting in the last 7 days. */
  topCommenters: CommenterUserConnection
  /** Returns the user with the username. */
  user?: Maybe<User>
}
// ...
export type GetBlogPostsQueryVariables = Exact<{ [key: string]: never }>

export type GetBlogPostsQuery = {
  __typename?: 'Query'
  publication?: {
    __typename?: 'Publication'
    posts: {
      __typename?: 'PublicationPostConnection'
      edges: Array<{
        __typename?: 'PostEdge'
        node: {
          __typename?: 'Post'
          id: string
          title: string
          url: string
          coverImage?: { __typename?: 'PostCoverImage'; url: string } | null
        }
      }>
    }
  } | null
}

export const GetBlogPostsDocument = {
  kind: 'Document',
  definitions: [
    {
      kind: 'OperationDefinition',
      operation: 'query',
      name: { kind: 'Name', value: 'getBlogPosts' },
      selectionSet: {
        kind: 'SelectionSet',
        selections: [
          {
            kind: 'Field',
            name: { kind: 'Name', value: 'publication' },
            arguments: [
              {
                kind: 'Argument',
                name: { kind: 'Name', value: 'host' },
                value: {
                  kind: 'StringValue',
                  value: 'luisfalconmx.hashnode.dev',
                  block: false
                }
              }
            ],
            selectionSet: {
              kind: 'SelectionSet',
              selections: [
                {
                  kind: 'Field',
                  name: { kind: 'Name', value: 'posts' },
                  arguments: [
                    {
                      kind: 'Argument',
                      name: { kind: 'Name', value: 'first' },
                      value: { kind: 'IntValue', value: '10' }
                    }
                  ],
                  selectionSet: {
                    kind: 'SelectionSet',
                    selections: [
                      {
                        kind: 'Field',
                        name: { kind: 'Name', value: 'edges' },
                        selectionSet: {
                          kind: 'SelectionSet',
                          selections: [
                            {
                              kind: 'Field',
                              name: { kind: 'Name', value: 'node' },
                              selectionSet: {
                                kind: 'SelectionSet',
                                selections: [
                                  {
                                    kind: 'Field',
                                    name: { kind: 'Name', value: 'id' }
                                  },
                                  {
                                    kind: 'Field',
                                    name: { kind: 'Name', value: 'title' }
                                  },
                                  {
                                    kind: 'Field',
                                    name: { kind: 'Name', value: 'url' }
                                  },
                                  {
                                    kind: 'Field',
                                    name: { kind: 'Name', value: 'coverImage' },
                                    selectionSet: {
                                      kind: 'SelectionSet',
                                      selections: [
                                        {
                                          kind: 'Field',
                                          name: { kind: 'Name', value: 'url' }
                                        }
                                      ]
                                    }
                                  }
                                ]
                              }
                            }
                          ]
                        }
                      }
                    ]
                  }
                }
              ]
            }
          }
        ]
      }
    }
  ]
} as unknown as DocumentNode<GetBlogPostsQuery, GetBlogPostsQueryVariables>